requirements.txt
altair==4.1.0
altgraph==0.17
anaconda==0.0.1.1
asgiref==3.2.10
asn1crypto==1.4.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
boto3==1.16.57
botocore==1.19.57
bs4==0.0.1
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
client==0.0.1
colorama==0.4.3
cryptography==2.9.2
cssselect==1.1.0
cssutils==1.0.2
cx-Freeze==6.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.7
docopt==0.6.2
docutils==0.16
emails==0.5.15
entrypoints==0.3
enum-compat==0.0.3
filesplit==3.0.2
fileupload==0.1.5
Flask==1.1.1
future==0.18.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
h5py==2.10.0
httplib2==0.18.1
idna==2.8
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonpickle==1.4.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.2.0
localStoragePy==0.2.3
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
more-itertools==8.4.0
mysql-connector==2.2.9
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.7
networks==0.3.7
notebook==6.0.3
numpy==1.14.3
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.2.2
pipreqs==0.4.10
plyer==1.4.3
pocketsphinx==0.1.15
premailer==3.7.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pycryptodome==3.10.1
pydeck==0.2.1
Pygments==2.6.1
PyInstaller==3.6
pyler==0.2.0
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qrcode==6.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.4
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
sip==5.3.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
streamlit==0.56.0
terminado==0.8.3
testpath==0.4.4
Theano==1.0.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.11.1
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
validators==0.18.2
watchdog==1.0.2
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
wordcloud @ file:///C:/Users/pc/Desktop/wordcloud-1.7.0-cp36-cp36m-win32.whl
yarg==0.1.9
zipp==3.1.0
